Show RequestBody form only when the HTTP method is PUT or POST in DocService Motivation:
Currently, documentation service shows request body form regardless of the HTTP method.
We should show it only when the HTTP method is PUT or POST.

Modifications:

- Show request body from only when the HTTP method is POST or PUT.
- Split `MethodPage` into several `SFC`s and `DebugPage`.
  - All the states in `MethodPage` are now moved into `DebugPage`.
- Fix a bug where uniqe key is not assigned when creating a type link.

Result:

- The request body form is no longer shown when the HTTP method is not POST or PUT.

Todos:

- More refactoring (divide `DebugPage` into `RpcDebugPage` and `HttpDebugPage`?)
