+++ docs-client/src/containers/App/index.tsx
@@ -46,7 +46,6 @@ import MethodPage from '../MethodPage';
 import StructPage from '../StructPage';
 
 import {
-  methodKey,
   simpleName,
   Specification,
   SpecificationData,
@@ -188,19 +187,15 @@ function AppDrawer({
                   {service.methods.map((method) => (
                     <ListItem
                       dense
-                      key={methodKey(
-                        service.name,
-                        method.name,
-                        method.httpMethod,
-                      )}
+                      key={`${service.name}/${method.name}/${
+                        method.httpMethod
+                      }`}
                       button
                       onClick={() =>
                         navigateTo(
-                          `/methods/${methodKey(
-                            service.name,
-                            method.name,
-                            method.httpMethod,
-                          )}`,
+                          `/methods/${service.name}/${method.name}/${
+                            method.httpMethod
+                          }`,
                         )
                       }
                     >
@@ -434,7 +429,7 @@ class App extends React.PureComponent<Props, State> {
             )}
           />
           <Route
-            path="/methods/:serviceName/:methodName"
+            path="/methods/:serviceName/:methodName/:httpMethod"
             render={(props) => (
               <MethodPage {...props} specification={specification} />
             )}
